Move all the iloc code from column to series for better host buffering.
